Use Float for BPM display
A cosmetic change on how the BPM display works: it visualizes what `Infinity` really does to the math, alongside showing how accurate the BPM calculation is per map.

Does not affect the actual timings calculation in any way.

DB modified to reflect this change while retaining a compatible layout for previous versions.

+++ src/itdelatrisu/opsu/beatmap/Beatmap.java
@@ -217,7 +217,7 @@ public class Beatmap implements Comparable<Beatmap> {
 	public ArrayList<TimingPoint> timingPoints;
 
 	/** Song BPM range. */
-	public int bpmMin = 0, bpmMax = 0;
+	public float bpmMin = 0, bpmMax = 0;
 
 	/**
 	 * [Colours]
